@@ -1,7 +1,7 @@
 plugins {
     id 'java'
     id 'maven-publish'
-    id "com.github.johnrengelman.shadow" version "4.0.4"
+    id "com.github.johnrengelman.shadow" version "7.0.0"
 }
 
 defaultTasks 'build'
@@ -12,6 +12,7 @@ description 'A lightweight yet very powerful hologram plugin with many features 
 
 repositories {
     mavenCentral()
+    maven { url = "https://repo.papermc.io/repository/maven-public/"}
     maven { url = 'https://oss.sonatype.org/content/repositories/snapshots' }
     maven { url = 'https://repo.codemc.io/repository/nms/' }
     maven { url = 'https://repo.extendedclip.com/content/repositories/placeholderapi/' }
@@ -19,6 +20,7 @@ repositories {
 }
 
 dependencies {
+    compileOnly 'dev.folia:folia-api:1.19.4-R0.1-SNAPSHOT'
     compileOnly 'org.spigotmc:spigot:1.13-R0.1-SNAPSHOT'
     compileOnly 'me.clip:placeholderapi:2.11.2'
     compileOnly 'com.arcaniax:HeadDatabase-API:1.3.1'
@@ -32,8 +34,8 @@ dependencies {
     annotationProcessor 'org.jetbrains:annotations:24.0.0'
 }
 
-sourceCompatibility = 1.8
-targetCompatibility = 1.8
+sourceCompatibility = JavaVersion.VERSION_17
+targetCompatibility = JavaVersion.VERSION_17
 
 compileJava {
     options.encoding = 'UTF-8'
